backend/requirements.txt
Django
djangorestframework
pillow
pytesseract
opencv-python-headless
rapidfuzz
spacy
scikit-learn
pandas
python-multipart
